+++ Settings/KeybindPatch.cs
@@ -20,12 +20,13 @@ namespace PerformanceSettings.Settings
             {
                 if (data.cantUseItemFor <= 0f)
                 {
-                    KeyCode key1 = (KeyCode)GameHandler.Instance.SettingsHandler.GetSetting<UseItemKeybindSetting>().Value;
+                    
+                    KeyCode key1 = (KeyCode)ContentSettings.API.SettingsLoader.GetSetting<UseItemKeybindSetting>()!.Value;
                     __instance.clickWasPressed = GlobalInputHandler.GetKeyDown(key1);
                     __instance.clickIsPressed = GlobalInputHandler.GetKey(key1);
                     __instance.clickWasReleased = GlobalInputHandler.GetKeyUp(key1);
                 }
-                KeyCode key2 = (KeyCode)GameHandler.Instance.SettingsHandler.GetSetting<AimKeybindSetting>().Value;
+                KeyCode key2 = (KeyCode)ContentSettings.API.SettingsLoader.GetSetting<AimKeybindSetting>()!.Value;
                 __instance.aimWasPressed = GlobalInputHandler.GetKeyDown(key2);
                 __instance.aimIsPressed = GlobalInputHandler.GetKey(key2);
             }